requirements.txt
Flask
python-dotenv
Flask-Session
Spotipy